+++ src/main/java/frc/robot/commands/scoring/intake/IntakeGamePiece.java
@@ -8,13 +8,13 @@ import edu.wpi.first.wpilibj.DriverStation;
 import edu.wpi.first.wpilibj.Timer;
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.constants.IntakeConstants;
-import frc.robot.subsystems.RollerIntake;
-import frc.robot.subsystems.RollerIntake.IntakeMode;
+import frc.robot.subsystems.Intake;
+import frc.robot.subsystems.Intake.IntakeMode;
 import frc.robot.util.GamePieceType;
 
 public class IntakeGamePiece extends CommandBase {
 
-  private final RollerIntake m_intake; 
+  private final Intake m_intake; 
   private GamePieceType m_type;
   private final BooleanSupplier m_isCone;
   private final boolean m_runsForever;
@@ -26,7 +26,7 @@ public class IntakeGamePiece extends CommandBase {
    * @param isCone a supplier that when the command starts, checks if will intake a cone or cube
    * @param runsForever if the command should end, or run forever. If false, will end based on motor currents
    */
-  public IntakeGamePiece(RollerIntake intake, BooleanSupplier isCone, boolean runsForever) {
+  public IntakeGamePiece(Intake intake, BooleanSupplier isCone, boolean runsForever) {
     m_intake = intake;
     m_isCone = isCone;
     m_runsForever = runsForever;
