+++ src/main/java/frc/robot/subsystems/Intake.java
@@ -23,7 +23,7 @@ import frc.robot.util.LogManager;
 import frc.robot.util.MotorFactory;
 
 
-public class RollerIntake extends SubsystemBase {
+public class Intake extends SubsystemBase {
 
   public enum IntakeMode {
     INTAKE_CUBE, OUTTAKE_CUBE, INTAKE_CONE, OUTTAKE_CONE, DISABLED
@@ -39,7 +39,7 @@ public class RollerIntake extends SubsystemBase {
   private double m_power;
 
 
-  public RollerIntake(ShuffleboardTab intakeTab) {
+  public Intake(ShuffleboardTab intakeTab) {
     m_intakeMotor = MotorFactory.createTalonFX(IntakeConstants.kIntakeMotorId, Constants.kRioCAN);
     m_intakeMotor.setNeutralMode(IntakeConstants.kNeutralMode);
     m_intakeMotor.enableVoltageCompensation(true);
